Fix all the format string warnings. Some other random warnings.
Acked-by: Stefan Reinauer <stepan@coresystems.de>


git-svn-id: svn://svn.coreboot.org/coreboot/trunk@5268 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/southbridge/intel/i3100/i3100_lpc.c
@@ -317,7 +317,7 @@ static void lpc_init(struct device *dev)
 	if (!res) {
 		return;
 	}
-	*((u8 *)(res->base + 0x31ff)) |= (1 << 0);
+	*((u8 *)((u32)res->base + 0x31ff)) |= (1 << 0);
 
 	// TODO this code sets int 0 of the IOAPIC in Virtual Wire Mode
 	// (register 0x10/0x11) while the old code used int 1 (register 0x12)
